+++ backend/domain_types/src/types.rs
@@ -60,9 +60,8 @@ fn extract_headers_from_metadata(
 // For decoding connector_meta_data and Engine trait - base64 crate no longer needed here
 use crate::{
     connector_flow::{
-        Accept, Authenticate, Authorize, Capture, CreateOrder, CreateSessionToken, DefendDispute,
-        PSync, PaymentMethodToken, PostAuthenticate, PreAuthenticate, RSync, Refund, RepeatPayment,
-        SetupMandate, SubmitEvidence, Void,
+        Accept, Authorize, Capture, CreateOrder, CreateSessionToken, DefendDispute, PSync,
+        PaymentMethodToken, RSync, Refund, RepeatPayment, SetupMandate, SubmitEvidence, Void,
     },
     connector_types::{
         AcceptDisputeData, AccessTokenRequestData, ConnectorMandateReferenceId,
@@ -4905,598 +4904,6 @@ pub fn generate_repeat_payment_response(
             )
         }
     }
-}
-
-pub fn generate_payment_pre_authenticate_response<T: PaymentMethodDataTypes>(
-    router_data_v2: RouterDataV2<
-        PreAuthenticate,
-        PaymentFlowData,
-        PaymentsPreAuthenticateData<T>,
-        PaymentsResponseData,
-    >,
-) -> Result<
-    grpc_api_types::payments::PaymentServicePreAuthenticateResponse,
-    error_stack::Report<ApplicationErrorResponse>,
-> {
-    let transaction_response = router_data_v2.response;
-    let status = router_data_v2.resource_common_data.status;
-    let grpc_status = grpc_api_types::payments::PaymentStatus::foreign_from(status);
-    let raw_connector_response = router_data_v2
-        .resource_common_data
-        .get_raw_connector_response();
-    let response_headers = router_data_v2
-        .resource_common_data
-        .get_connector_response_headers_as_map();
-
-    let response = match transaction_response {
-        Ok(response) => match response {
-            PaymentsResponseData::PreAuthenticateResponse {
-                resource_id,
-                redirection_data,
-                connector_metadata,
-                connector_response_reference_id,
-                status_code,
-            } => grpc_api_types::payments::PaymentServicePreAuthenticateResponse {
-                transaction_id: Some(grpc_api_types::payments::Identifier::foreign_try_from(
-                    resource_id,
-                )?),
-                redirection_data: redirection_data
-                    .map(|form| match *form {
-                        router_response_types::RedirectForm::Form {
-                            endpoint,
-                            method,
-                            form_fields,
-                        } => Ok::<grpc_api_types::payments::RedirectForm, ApplicationErrorResponse>(
-                            grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Form(
-                                        grpc_api_types::payments::FormData {
-                                            endpoint,
-                                            method:
-                                                grpc_api_types::payments::HttpMethod::foreign_from(
-                                                    method,
-                                                )
-                                                    as i32,
-                                            form_fields,
-                                        },
-                                    ),
-                                ),
-                            },
-                        ),
-                        router_response_types::RedirectForm::Html { html_data } => {
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Html(
-                                        grpc_api_types::payments::HtmlData { html_data },
-                                    ),
-                                ),
-                            })
-                        }
-                        router_response_types::RedirectForm::Uri { uri } => {
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Uri(
-                                        grpc_api_types::payments::UriData { uri },
-                                    ),
-                                ),
-                            })
-                        }
-                        router_response_types::RedirectForm::Mifinity {
-                            initialization_token,
-                        } => Ok(grpc_api_types::payments::RedirectForm {
-                            form_type: Some(
-                                grpc_api_types::payments::redirect_form::FormType::Uri(
-                                    grpc_api_types::payments::UriData {
-                                        uri: initialization_token,
-                                    },
-                                ),
-                            ),
-                        }),
-                        router_response_types::RedirectForm::CybersourceAuthSetup {
-                            access_token,
-                            ddc_url,
-                            reference_id,
-                        } => {
-                            let mut form_fields = std::collections::HashMap::new();
-                            form_fields.insert("access_token".to_string(), access_token);
-                            form_fields.insert("ddc_url".to_string(), ddc_url.clone());
-                            form_fields.insert("reference_id".to_string(), reference_id);
-
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Form(
-                                        grpc_api_types::payments::FormData {
-                                            endpoint: ddc_url,
-                                            method: grpc_api_types::payments::HttpMethod::Post
-                                                as i32,
-                                            form_fields,
-                                        },
-                                    ),
-                                ),
-                            })
-                        }
-                        _ => Err(ApplicationErrorResponse::BadRequest(ApiError {
-                            sub_code: "INVALID_RESPONSE".to_owned(),
-                            error_identifier: 400,
-                            error_message: "Invalid response from connector".to_owned(),
-                            error_object: None,
-                        }))?,
-                    })
-                    .transpose()?,
-                connector_metadata: connector_metadata
-                    .and_then(|value| value.as_object().cloned())
-                    .map(|map| {
-                        map.into_iter()
-                            .filter_map(|(k, v)| v.as_str().map(|s| (k, s.to_string())))
-                            .collect::<HashMap<_, _>>()
-                    })
-                    .unwrap_or_default(),
-                response_ref_id: connector_response_reference_id.map(|id| {
-                    grpc_api_types::payments::Identifier {
-                        id_type: Some(grpc_api_types::payments::identifier::IdType::Id(id)),
-                    }
-                }),
-                status: grpc_status as i32,
-                error_message: None,
-                error_code: None,
-                raw_connector_response,
-                status_code: status_code as u32,
-                response_headers,
-                network_txn_id: None,
-                state: None,
-            },
-            _ => {
-                return Err(ApplicationErrorResponse::BadRequest(ApiError {
-                    sub_code: "INVALID_RESPONSE".to_owned(),
-                    error_identifier: 400,
-                    error_message: "Invalid response type for pre authenticate".to_owned(),
-                    error_object: None,
-                })
-                .into())
-            }
-        },
-        Err(err) => {
-            let status = err
-                .attempt_status
-                .map(grpc_api_types::payments::PaymentStatus::foreign_from)
-                .unwrap_or_default();
-            grpc_api_types::payments::PaymentServicePreAuthenticateResponse {
-                transaction_id: Some(grpc_api_types::payments::Identifier {
-                    id_type: Some(
-                        grpc_api_types::payments::identifier::IdType::NoResponseIdMarker(()),
-                    ),
-                }),
-                redirection_data: None,
-                network_txn_id: None,
-                response_ref_id: None,
-                status: status as i32,
-                error_message: Some(err.message),
-                error_code: Some(err.code),
-                status_code: err.status_code as u32,
-                response_headers,
-                raw_connector_response,
-                connector_metadata: std::collections::HashMap::new(),
-                state: None,
-            }
-        }
-    };
-    Ok(response)
-}
-
-pub fn generate_payment_authenticate_response<T: PaymentMethodDataTypes>(
-    router_data_v2: RouterDataV2<
-        Authenticate,
-        PaymentFlowData,
-        PaymentsAuthenticateData<T>,
-        PaymentsResponseData,
-    >,
-) -> Result<
-    grpc_api_types::payments::PaymentServiceAuthenticateResponse,
-    error_stack::Report<ApplicationErrorResponse>,
-> {
-    let transaction_response = router_data_v2.response;
-    let status = router_data_v2.resource_common_data.status;
-    let grpc_status = grpc_api_types::payments::PaymentStatus::foreign_from(status);
-    let raw_connector_response = router_data_v2
-        .resource_common_data
-        .get_raw_connector_response();
-    let response_headers = router_data_v2
-        .resource_common_data
-        .get_connector_response_headers_as_map();
-
-    let response = match transaction_response {
-        Ok(response) => match response {
-            PaymentsResponseData::AuthenticateResponse {
-                resource_id,
-                redirection_data,
-                connector_metadata,
-                connector_response_reference_id,
-                status_code,
-            } => grpc_api_types::payments::PaymentServiceAuthenticateResponse {
-                transaction_id: Some(grpc_api_types::payments::Identifier::foreign_try_from(
-                    resource_id,
-                )?),
-                redirection_data: redirection_data
-                    .map(|form| match *form {
-                        router_response_types::RedirectForm::Form {
-                            endpoint,
-                            method,
-                            form_fields,
-                        } => Ok::<grpc_api_types::payments::RedirectForm, ApplicationErrorResponse>(
-                            grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Form(
-                                        grpc_api_types::payments::FormData {
-                                            endpoint,
-                                            method:
-                                                grpc_api_types::payments::HttpMethod::foreign_from(
-                                                    method,
-                                                )
-                                                    as i32,
-                                            form_fields,
-                                        },
-                                    ),
-                                ),
-                            },
-                        ),
-                        router_response_types::RedirectForm::Html { html_data } => {
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Html(
-                                        grpc_api_types::payments::HtmlData { html_data },
-                                    ),
-                                ),
-                            })
-                        }
-                        router_response_types::RedirectForm::Uri { uri } => {
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Uri(
-                                        grpc_api_types::payments::UriData { uri },
-                                    ),
-                                ),
-                            })
-                        }
-                        router_response_types::RedirectForm::Mifinity {
-                            initialization_token,
-                        } => Ok(grpc_api_types::payments::RedirectForm {
-                            form_type: Some(
-                                grpc_api_types::payments::redirect_form::FormType::Uri(
-                                    grpc_api_types::payments::UriData {
-                                        uri: initialization_token,
-                                    },
-                                ),
-                            ),
-                        }),
-                        router_response_types::RedirectForm::CybersourceAuthSetup {
-                            access_token,
-                            ddc_url,
-                            reference_id,
-                        } => {
-                            let mut form_fields = std::collections::HashMap::new();
-                            form_fields.insert("access_token".to_string(), access_token);
-                            form_fields.insert("ddc_url".to_string(), ddc_url.clone());
-                            form_fields.insert("reference_id".to_string(), reference_id);
-
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Form(
-                                        grpc_api_types::payments::FormData {
-                                            endpoint: ddc_url,
-                                            method: grpc_api_types::payments::HttpMethod::Post
-                                                as i32,
-                                            form_fields,
-                                        },
-                                    ),
-                                ),
-                            })
-                        }
-                        _ => Err(ApplicationErrorResponse::BadRequest(ApiError {
-                            sub_code: "INVALID_RESPONSE".to_owned(),
-                            error_identifier: 400,
-                            error_message: "Invalid response from connector".to_owned(),
-                            error_object: None,
-                        }))?,
-                    })
-                    .transpose()?,
-                connector_metadata: connector_metadata
-                    .as_ref()
-                    .and_then(|value| value.as_object().cloned())
-                    .map(|map| {
-                        map.into_iter()
-                            .filter_map(|(k, v)| v.as_str().map(|s| (k, s.to_string())))
-                            .collect::<HashMap<_, _>>()
-                    })
-                    .unwrap_or_default(),
-                response_ref_id: connector_response_reference_id.map(|id| {
-                    grpc_api_types::payments::Identifier {
-                        id_type: Some(grpc_api_types::payments::identifier::IdType::Id(id)),
-                    }
-                }),
-                authentication_data: connector_metadata.as_ref().map(|metadata| {
-                    grpc_api_types::payments::AuthenticationData {
-                        eci: metadata
-                            .get("eci")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string()),
-                        cavv: metadata
-                            .get("cavv")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string())
-                            .unwrap_or_default(),
-                        threeds_server_transaction_id: metadata
-                            .get("threeds_server_transaction_id")
-                            .and_then(|v| v.as_str())
-                            .map(|s| grpc_api_types::payments::Identifier {
-                                id_type: Some(grpc_api_types::payments::identifier::IdType::Id(
-                                    s.to_string(),
-                                )),
-                            }),
-                        message_version: metadata
-                            .get("message_version")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string()),
-                        ds_transaction_id: metadata
-                            .get("ds_transaction_id")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string()),
-                    }
-                }),
-                status: grpc_status as i32,
-                error_message: None,
-                error_code: None,
-                raw_connector_response,
-                status_code: status_code as u32,
-                response_headers,
-                network_txn_id: None,
-                state: None,
-            },
-            _ => {
-                return Err(ApplicationErrorResponse::BadRequest(ApiError {
-                    sub_code: "INVALID_RESPONSE".to_owned(),
-                    error_identifier: 400,
-                    error_message: "Invalid response type for authenticate".to_owned(),
-                    error_object: None,
-                })
-                .into())
-            }
-        },
-        Err(err) => {
-            let status = err
-                .attempt_status
-                .map(grpc_api_types::payments::PaymentStatus::foreign_from)
-                .unwrap_or_default();
-            grpc_api_types::payments::PaymentServiceAuthenticateResponse {
-                transaction_id: Some(grpc_api_types::payments::Identifier {
-                    id_type: Some(grpc_api_types::payments::identifier::IdType::Id(
-                        "session_created".to_string(),
-                    )),
-                }),
-                redirection_data: None,
-                network_txn_id: None,
-                response_ref_id: None,
-                authentication_data: None,
-                status: status as i32,
-                error_message: Some(err.message),
-                error_code: Some(err.code),
-                status_code: err.status_code as u32,
-                raw_connector_response,
-                response_headers,
-                connector_metadata: std::collections::HashMap::new(),
-                state: None,
-            }
-        }
-    };
-    Ok(response)
-}
-
-pub fn generate_payment_post_authenticate_response<T: PaymentMethodDataTypes>(
-    router_data_v2: RouterDataV2<
-        PostAuthenticate,
-        PaymentFlowData,
-        PaymentsPostAuthenticateData<T>,
-        PaymentsResponseData,
-    >,
-) -> Result<
-    grpc_api_types::payments::PaymentServicePostAuthenticateResponse,
-    error_stack::Report<ApplicationErrorResponse>,
-> {
-    let transaction_response = router_data_v2.response;
-    let status = router_data_v2.resource_common_data.status;
-    let grpc_status = grpc_api_types::payments::PaymentStatus::foreign_from(status);
-    let raw_connector_response = router_data_v2
-        .resource_common_data
-        .get_raw_connector_response();
-    let response_headers = router_data_v2
-        .resource_common_data
-        .get_connector_response_headers_as_map();
-
-    let response = match transaction_response {
-        Ok(response) => match response {
-            PaymentsResponseData::PostAuthenticateResponse {
-                resource_id,
-                redirection_data,
-                connector_metadata,
-                connector_response_reference_id,
-                status_code,
-            } => grpc_api_types::payments::PaymentServicePostAuthenticateResponse {
-                transaction_id: Some(grpc_api_types::payments::Identifier::foreign_try_from(
-                    resource_id,
-                )?),
-                redirection_data: redirection_data
-                    .map(|form| match *form {
-                        router_response_types::RedirectForm::Form {
-                            endpoint,
-                            method,
-                            form_fields,
-                        } => Ok::<grpc_api_types::payments::RedirectForm, ApplicationErrorResponse>(
-                            grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Form(
-                                        grpc_api_types::payments::FormData {
-                                            endpoint,
-                                            method:
-                                                grpc_api_types::payments::HttpMethod::foreign_from(
-                                                    method,
-                                                )
-                                                    as i32,
-                                            form_fields,
-                                        },
-                                    ),
-                                ),
-                            },
-                        ),
-                        router_response_types::RedirectForm::Html { html_data } => {
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Html(
-                                        grpc_api_types::payments::HtmlData { html_data },
-                                    ),
-                                ),
-                            })
-                        }
-                        router_response_types::RedirectForm::Uri { uri } => {
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Uri(
-                                        grpc_api_types::payments::UriData { uri },
-                                    ),
-                                ),
-                            })
-                        }
-                        router_response_types::RedirectForm::Mifinity {
-                            initialization_token,
-                        } => Ok(grpc_api_types::payments::RedirectForm {
-                            form_type: Some(
-                                grpc_api_types::payments::redirect_form::FormType::Uri(
-                                    grpc_api_types::payments::UriData {
-                                        uri: initialization_token,
-                                    },
-                                ),
-                            ),
-                        }),
-                        router_response_types::RedirectForm::CybersourceAuthSetup {
-                            access_token,
-                            ddc_url,
-                            reference_id,
-                        } => {
-                            let mut form_fields = std::collections::HashMap::new();
-                            form_fields.insert("access_token".to_string(), access_token);
-                            form_fields.insert("ddc_url".to_string(), ddc_url.clone());
-                            form_fields.insert("reference_id".to_string(), reference_id);
-
-                            Ok(grpc_api_types::payments::RedirectForm {
-                                form_type: Some(
-                                    grpc_api_types::payments::redirect_form::FormType::Form(
-                                        grpc_api_types::payments::FormData {
-                                            endpoint: ddc_url,
-                                            method: grpc_api_types::payments::HttpMethod::Post
-                                                as i32,
-                                            form_fields,
-                                        },
-                                    ),
-                                ),
-                            })
-                        }
-                        _ => Err(ApplicationErrorResponse::BadRequest(ApiError {
-                            sub_code: "INVALID_RESPONSE".to_owned(),
-                            error_identifier: 400,
-                            error_message: "Invalid response from connector".to_owned(),
-                            error_object: None,
-                        }))?,
-                    })
-                    .transpose()?,
-                connector_metadata: connector_metadata
-                    .as_ref()
-                    .and_then(|value| value.as_object().cloned())
-                    .map(|map| {
-                        map.into_iter()
-                            .filter_map(|(k, v)| v.as_str().map(|s| (k, s.to_string())))
-                            .collect::<HashMap<_, _>>()
-                    })
-                    .unwrap_or_default(),
-                network_txn_id: None,
-                response_ref_id: connector_response_reference_id.map(|id| {
-                    grpc_api_types::payments::Identifier {
-                        id_type: Some(grpc_api_types::payments::identifier::IdType::Id(id)),
-                    }
-                }),
-                authentication_data: connector_metadata.as_ref().map(|metadata| {
-                    grpc_api_types::payments::AuthenticationData {
-                        eci: metadata
-                            .get("eci")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string()),
-                        cavv: metadata
-                            .get("cavv")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string())
-                            .unwrap_or_default(),
-                        threeds_server_transaction_id: metadata
-                            .get("threeds_server_transaction_id")
-                            .and_then(|v| v.as_str())
-                            .map(|s| grpc_api_types::payments::Identifier {
-                                id_type: Some(grpc_api_types::payments::identifier::IdType::Id(
-                                    s.to_string(),
-                                )),
-                            }),
-                        message_version: metadata
-                            .get("message_version")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string()),
-                        ds_transaction_id: metadata
-                            .get("ds_transaction_id")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string()),
-                    }
-                }),
-                incremental_authorization_allowed: None,
-                status: grpc_status as i32,
-                error_message: None,
-                error_code: None,
-                raw_connector_response,
-                status_code: status_code as u32,
-                response_headers,
-                state: None,
-            },
-            _ => {
-                return Err(ApplicationErrorResponse::BadRequest(ApiError {
-                    sub_code: "INVALID_RESPONSE".to_owned(),
-                    error_identifier: 400,
-                    error_message: "Invalid response type for post authenticate".to_owned(),
-                    error_object: None,
-                })
-                .into())
-            }
-        },
-        Err(err) => {
-            let status = err
-                .attempt_status
-                .map(grpc_api_types::payments::PaymentStatus::foreign_from)
-                .unwrap_or_default();
-            grpc_api_types::payments::PaymentServicePostAuthenticateResponse {
-                transaction_id: Some(grpc_api_types::payments::Identifier {
-                    id_type: Some(
-                        grpc_api_types::payments::identifier::IdType::NoResponseIdMarker(()),
-                    ),
-                }),
-                redirection_data: None,
-                network_txn_id: None,
-                response_ref_id: None,
-                authentication_data: None,
-                incremental_authorization_allowed: None,
-                status: status as i32,
-                error_message: Some(err.message),
-                error_code: Some(err.code),
-                status_code: err.status_code as u32,
-                response_headers,
-                raw_connector_response,
-                connector_metadata: std::collections::HashMap::new(),
-                state: None,
-            }
-        }
-    };
-    Ok(response)
 }
 
 // New ForeignTryFrom implementations for individual 3DS authentication flow proto definitions
